requirements.txt
requests
colorama
pytchat
openai
python-dotenv
soundfile
sounddevice
pyttsx3